Refactor to cookie_repository
